Move `pip._internal:main` to its own module
Moving content out of `__init__` is preferred in general because it
avoids conflicts with module names and unnecessary imports.
